pyproject.toml
[tool.poetry]
name = "simple-chat-isi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "simple_chat_isi"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
django-jazzmin = "^2.6.0"
psycopg2-binary = "^2.9.6"
django-split-settings = "^1.2.0"
django-cors-headers = "^3.14.0"
django-autocomplete-light = "^3.9.4"
drf-spectacular = "^0.26.1"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
